Add preliminary files for package
This commit begins the process of porting the big Pluto notebook to a
proper Julia package.  I've made a few changes from the notebook, so
there's a chance that the code won't work as-is.  There are also no
tests yet, so definitely don't try to use it (or anything in this
branch!) for serious usage.