Remove L, A, and nz fields from Automaton
They're so easy to compute from the other fields that storing them
seems like a minor waste of memory.  This commit replaces the old L and
A fields with functions nlocations(a) and nactions(a).  It implements
Base.getproperty to provide the nz field, as well as new L and A fields
that are Base.oneto ranges, better matching the semantics of the paper.