Change default "hit" action to 1
It makes more sense to keep the default action as 1, since then we could
have even a trivial automaton with one action and the default still
applies.  Since a deadline hit should be the expected behavior, this
commit changes the Automaton constructors so a hit is 1 and a miss is 2.