Add human-readable list of Automaton constructors
For presenting the different deadline miss handling strategies in user
interfaces, such as Pluto notebooks, it is very useful to have a map of
human-readable names to the constructor functions.  Since this will
likely be used more than once, it makes sense to add it to the package
rather than making each notebook redefine it on its own, perhaps
inconsistently.  This commit adds such a map, as well as a
lexicographically sorted list of the names.