Add first half of reimplemented notebook
The original control_deadline_misses_automata.jl notebook contained the
code used for the paper, but the whole point of this repository is to
refactor that code into a reusable Julia package.  Given this goal, it
makes sense to rewrite the notebook to actually use the package, rather
than defining everything on its own.  This commit adds a new notebook,
control_timing_safety.jl, that will reach this aim.  For now, it
contains just the first section (with some significant improvements in
Example 2).  The rest will follow later.