Add support for vectors as bounds
Previously, we required that the initial set be given as a bounding box
for all the major API functions.  This resulted in some inefficiency
when the user only wanted a single initial state, in particular for the
`bounded_runs` function.  This commit adds support for `bounds` being a
vector instead of a matrix to alleviate this computational overhead.

Closes #4.