Remove Julia 1.7 array syntax, use ones
Without thinking, I added some array literals using the Julia 1.7 syntax
of multiple semicolons in a row.  Since this was just building matrices
from column vectors, simply using a space would do the job, without
breaking support for the current LTS Julia 1.6.

The default value of nominal in the deviation function used repeat.  It
was just filling a vector with ones, so it's better to use the ones
function for that.