Use MissRow constraints from RealTimeScheduling.jl
Now that RealTimeScheduling.jl has a real release, we can depend on it
in this package without requiring manual dependency management or
something.  This allows us to use a WeaklyHardConstraint to specify how
an Automaton should be constructed, paving the way to a clean API when
support for more types of constraint is added.

Closes #6.