Support deviation from evol output
Previously, the `deviation` function assumed its parameter `reachable`
was the output of `bounded_runs_iter`.  This would make an unnecessary
difficulty when testing deviation of two individual runs.  This commit
allows the parameter to be two-dimensional, so the output of `evol` can
be provided as `reachable`.  More API changes are still to come for this
function.