Implement JBF-based statistical deviation estimation (#9)
This commit adds a few new functions to estimate the maximum deviation that an Automaton may experience using Jeffreys' Bayes factor hypothesis testing. This is the method from "Statistical Hypothesis Testing of Controller Implementations Under Timing Uncertainties". For now, we only support a single initial point, but support for an initial set will be forthcoming.

Closes #7.