Allocate less memory in maximum_deviation_random
A lot of the time spent in maximum_deviation_random is due to memory
allocation and garbage collection.  Reducing this will probably improve
performance a lot, and make parallelizing the function more beneficial.
This commit makes a simple first pass at that, using in-place functions
for generating random behaviors and evolutions.  There are still many
allocations that would be good to remove, but this commit brings it
down enough to roughly halve the execution time in my testing.