Use output matrix C instead of dims
Previously, we selected the dimensions to be used when computing
deviation simply by specifying their indices.  This worked for initial
experiments, but is relatively inflexible.  The paper "Safety Analysis
of Embedded Controllers under Implementation Platform Timing
Uncertainties" specifies that the deviation is computed over the
system's *output*, which allows more flexibility.

This commit implements this by storing an output matrix as a property of
an Automaton object.  The miss handling strategy constructors take this
matrix as `sysd.C`.  The matrix is then used when computing deviation,
so it is no longer necessary to specify `dims`.