Further optimize bounded_runs memory usage
Originally, I had ordered the dimensions in a way that doesn't make
sense for Julia's column major arrays.  This commit rearranges arrays
for better cache behavior.  To avoid breaking users of the function, we
finally permute the dimensions to the old order, though this will likely
be removed in the future.