Add merge_bounds! function
Previously, we could only merge bounds with the allocating merge_bounds
function.  However, bounded_runs_iter already has memory allocated for
the merged bounds, so this allocation was redundant and wasteful.  This
commit adds a non-allocating version to enable more efficiency.  It also
rewrites the old one to use the non-allocating version under the hood,
which improves its efficiency substantially.