Reduce allocations in bounded_runs_iter
The function now uses the new merge_bounds! function, and pre-allocates
all the Automaton objects it needs instead of doing it on the fly more
times than necessary.  This only made a tiny dent in the total number of
allocations, though.  For the real improvement, I'll need to revisit
bounded_runs, and make a non-allocating version of *that*.