Hugely reduce allocations in bounded_runs
Due to how the stack was implemented, some operations like matrix
multiplication were allocating even when I used the non-allocating
versions of the functions.  This led to a lot of wasted memory
allocations.  This commit re-implements the stack with a mutable struct
for each frame, allowing us to use more non-allocating operations and
increasing the function's speed significantly.  This will also help with
making a non-allocating version of bounded_runs.