Add synthesize_constraints_deviation function
To support schedule optimization, we need to know the actual deviation
achieved by each safe constraint.  This commit adds a new function to
compute just that.

Tangentially, this commit also adds a parameter to
synthesize_constraints that specifies the automaton constructor to be
used, rather than hard-coding it as before.