Add heuristic for iteration order
Previously, we went over all elements of the product of constraints in
an arbitrary order, testing schedulability for all of them.  That was
woefully inefficient, running many schedulability tests for systems we
knew were dominated by others already.  Even worse when substituting in
the schedule_xghtc function for schedulable_jcl.  This commit adds a
heuristic, sorting all the constraints from smallest to largest
deviation, and skipping the test for any where the deviation is known
to be dominated.  This makes things much faster for JCLS, and actually
tractable for XGHTC.